# PERF: nancorr pearson