- [x] xref #40956
- [x] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing.html#code-standards) for how to run them
- [x] whatsnew entry

Benchmarks:
```
       before           after         ratio
     [e64fcfa7]       [c9a99585]
     <master>         <nancorr_perf>
            52.3M            52.1M     1.00  stat_ops.Correlation.peakmem_corr_wide('pearson')
-         307±5μs         217±10μs     0.71  stat_ops.Correlation.time_corr('pearson')
         132±10μs         138±10μs     1.05  stat_ops.Correlation.time_corr_series('pearson')
-      7.74±0.6ms       2.80±0.2ms     0.36  stat_ops.Correlation.time_corr_wide('pearson')
       9.04±0.7ms       8.22±0.8ms     0.91  stat_ops.Correlation.time_corr_wide_nans('pearson')
       3.92±0.2ms       4.00±0.3ms     1.02  stat_ops.Correlation.time_corrwith_cols('pearson')
       6.99±0.2ms       7.62±0.5ms     1.09  stat_ops.Correlation.time_corrwith_rows('pearson')
```